Implement session token distribution; shrink typeId to uint16_t
Session token (#4):
- game_serialize_config: append session token as final blob byte
- game_apply_config: extract token into optional sessionTokenOut param
- shareConfig: generate random token (1–255) via esp_random(),
  call radio.setSessionToken() before broadcast
- runWaiter: extract token after successful game_apply_config,
  call radio.setSessionToken() to lock in the session

typeId uint32_t → uint16_t:
- RadioPacket.typeId: 4 bytes → 2 bytes; RadioTypeId::UNIVERSAL updated
- LightAir_Game.typeId, _typeId member and all setTypeId/typeId() APIs
- LightAir_TotemDriver findRunner() signature and incomingTypeId local
- game_serialize_config/apply_config: read/write 2 bytes for typeId
- All ruleset typeId values: 0x00000001..5 → 0x0001..5
- RADIO_MAX_PAYLOAD 235 → 237; RADIO_OUT_PAYLOAD synced

Net per-packet saving: 2 bytes header + distributes session token.

+++ src/game/LightAir_Game.h
@@ -46,8 +46,8 @@ struct ScoreTable {
 // inheritance is required.
 //
 // Fields:
-//   typeId         — unique 32-bit game identifier.  Written as
-//                    the first 4 bytes of config blobs so receivers
+//   typeId         — unique 16-bit game identifier.  Written as
+//                    the first 2 bytes of config blobs so receivers
 //                    can verify compatibility before applying.
 //
 //   name           — short display name (≤15 chars) shown in the
@@ -132,7 +132,7 @@ struct ScoreTable {
 //   };
 //
 //   const LightAir_Game game_ffa = {
-//       .typeId         = 0x00000001,
+//       .typeId         = 0x0001,
 //       .name           = "Free for All",
 //       .configVars     = configVars,  .configCount    = 1,
 //       .monitorVars    = monitorVars, .monitorCount   = 2,
@@ -152,7 +152,7 @@ struct ScoreTable {
 //   }
 // ----------------------------------------------------------------
 struct LightAir_Game {
-    uint32_t             typeId;
+    uint16_t             typeId;
     const char*          name;
 
     const ConfigVar*     configVars;
